requirements.txt
discord
os
time
platform
datetime
humanfriendly
colorama
datetime
discord.py
discord.ext
discord.ui
typing
mysql.connector
random
json
easy_pil
requests
io
asyncio
string